requirements.txt
annotated-types==0.7.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==47.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.13
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
